+++ core_service.go
@@ -7,7 +7,7 @@ import (
 )
 
 const (
-	CurrentVersion = "1.2.0"
+	CurrentVersion = "1.3.0"
 )
 
 type CoreService struct {